be/doctor_service/doctors/tests.py
# Tệp kiểm thử cho ứng dụng doctors
from django.test import TestCase
from rest_framework.test import APIClient

class DoctorProfileTests(TestCase):
    def setUp(self):
        # Khởi tạo client để gửi yêu cầu API
        self.client = APIClient()

    def test_create_doctor_profile(self):
        # Viết test cases sau khi tích hợp xác thực
        pass

class DiagnosisTests(TestCase):
    def setUp(self):
        self.client = APIClient()

    def test_create_diagnosis(self):
        # Viết test cases sau
        pass