be/doctor_service/requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9
python-jose[cryptography]==3.3
gunicorn==20.1
requests==2.31  
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1